refactor(KERN-1560): extract generic AST utils to shared meerkat-core/utils
Move reusable, non-sql-to-meerkat-specific utilities to shared location:

- utils/ast-constants.ts: getConstantValue, getConstantTypeId, isNullConstant
- utils/ast-column-ref.ts: getQualifiedColumnRef, getColumnName, QualifiedColumnRef

These are generic DuckDB AST inspection functions usable by any module.
sql-to-meerkat/helpers/ast-utils.ts re-exports them for backward compat.

Kept in sql-to-meerkat: matchMeasureFromExpr, extractTableName,
hasRecursiveCteInMap, sanitizeForSerialize (domain-specific).

work-item: KERN-1560
